Use time position, not slider position

+++ nuclei-android/src/main/java/nuclei/media/MediaInterface.java
@@ -323,7 +323,7 @@ public class MediaInterface implements Destroyable {
         boolean onFastForward(MediaPlayerController controller, MediaControllerCompat.TransportControls controls);
         boolean onRewind(MediaPlayerController controller, MediaControllerCompat.TransportControls controls);
 
-        int getNextPosition(int nextPosition);
+        long getNextPosition(long nextPosition);
 
         void onLoading(MediaPlayerController controller);
